MB-49965 provide more granular errors for javascript udfs
jsevaluator can cast the error received by api methods to api.Error, and get an object.

Reviewed-on: https://review.couchbase.org/c/query/+/167258
Tested-by: Marco Greco <marco.greco@couchbase.com>

+++ errors/errors.go
@@ -143,7 +143,8 @@ func (e *err) Error() string {
 
 func (e *err) Object() map[string]interface{} {
 	m := map[string]interface{}{
-		"code":    e.ICode,
+		// only use standard data types in the object
+		"code":    int32(e.ICode),
 		"key":     e.IKey,
 		"message": e.InternalMsg,
 		"caller":  e.InternalCaller,
